requirements.txt
python-telegram-bot==21.6
pydantic==2.9.2
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.5
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pdfplumber==0.11.4
pypdf==5.0.1
rapidfuzz==3.9.7
python-dotenv==1.0.1
uvloop==0.20.0; sys_platform != 'win32'
